[FIX] #88 - 보드 섹션 헤더 셀에서 헤더 뷰로 변경
- 상단 고정되게 하기 위해서 셀이 아닌 섹션 헤더 뷰로 변경
