--- DooRiBon/DooRiBon/Sources/Base/Board/Cells/BoardHeaderTableViewCell.xib
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina5_9" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <customFonts key="customFonts">
-        <array key="SpoqaHanSansNeo-Bold.ttf">
-            <string>SpoqaHanSansNeo-Bold</string>
-        </array>
-        <array key="SpoqaHanSansNeo-Regular.ttf">
-            <string>SpoqaHanSansNeo-Regular</string>
-        </array>
-    </customFonts>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="BoardHeaderTableViewCell" rowHeight="52" id="KGk-i7-Jjw" customClass="BoardHeaderTableViewCell" customModule="DooRiBon" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="320" height="52"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="320" height="52"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="우리의 여행 목표" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Bex-Xz-BDc">
-                        <rect key="frame" x="18" y="14" width="106" height="23"/>
-                        <fontDescription key="fontDescription" name="SpoqaHanSansNeo-Bold" family="Spoqa Han Sans Neo" pointSize="16"/>
-                        <color key="textColor" name="black2"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Cod-mC-eHN">
-                        <rect key="frame" x="229" y="14.666666666666664" width="73" height="22"/>
-                        <color key="backgroundColor" name="pointOrange"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="22" id="HQ8-9x-IQ0"/>
-                            <constraint firstAttribute="width" constant="73" id="nIr-3x-HDj"/>
-                        </constraints>
-                        <fontDescription key="fontDescription" name="SpoqaHanSansNeo-Regular" family="Spoqa Han Sans Neo" pointSize="12"/>
-                        <state key="normal" title="+  추가하기">
-                            <color key="titleColor" name="white9"/>
-                        </state>
-                        <userDefinedRuntimeAttributes>
-                            <userDefinedRuntimeAttribute type="number" keyPath="cornerRadius">
-                                <real key="value" value="6"/>
-                            </userDefinedRuntimeAttribute>
-                        </userDefinedRuntimeAttributes>
-                        <connections>
-                            <action selector="writeButtonClicked:" destination="KGk-i7-Jjw" eventType="touchUpInside" id="20z-Cx-Zq9"/>
-                        </connections>
-                    </button>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="trailing" secondItem="Cod-mC-eHN" secondAttribute="trailing" constant="18" id="8cB-i9-kab"/>
-                    <constraint firstItem="Bex-Xz-BDc" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="18" id="OVA-Le-4pl"/>
-                    <constraint firstItem="Cod-mC-eHN" firstAttribute="centerY" secondItem="Bex-Xz-BDc" secondAttribute="centerY" id="Qo6-Rc-dM5"/>
-                    <constraint firstItem="Bex-Xz-BDc" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="14" id="hdy-YI-KBR"/>
-                    <constraint firstAttribute="bottom" secondItem="Bex-Xz-BDc" secondAttribute="bottom" constant="15" id="u7F-Zg-jir"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="subTitleLabel" destination="Bex-Xz-BDc" id="GED-tU-gqa"/>
-            </connections>
-            <point key="canvasLocation" x="137.59999999999999" y="141.13300492610838"/>
-        </tableViewCell>
-    </objects>
-    <resources>
-        <namedColor name="black2">
-            <color red="0.18000000715255737" green="0.19599999487400055" blue="0.21199999749660492" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-        <namedColor name="pointOrange">
-            <color red="1" green="0.48199999332427979" blue="0.31799998879432678" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-        <namedColor name="white9">
-            <color red="0.99199998378753662" green="0.99599999189376831" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-    </resources>
-</document>
